feat(ui-tests): avoid selecting dev tools window
